Upstream rebase (#29)
* configure_me_codegen retroactively reserved on our `bind_host` parame… (#520)

* configure_me_codegen retroactively reserved on our `bind_host` parameter name

* Add label and pray

* Add more labels why not

* Prepare 0.10.0 Release (#522)

* bump version

* CHANGELOG

* Ballista gets a docker image!!! (#521)

* Ballista gets a docker image!!!

* Enable flight sql

* Allow executing startup script

* Allow executing executables

* Clippy

* Remove capture group (#527)

* fix python build in CI (#528)

* fix python build in CI

* save progress

* use same min rust version in all crates

* fix

* use image from pyo3

* use newer image from pyo3

* do not require protoc

* wheels now generated

* rat - exclude generated file

* Update docs for simplified instructions (#532)

* Update docs for simplified instructions

* Fix whoopsie

* Update docs/source/user-guide/flightsql.md



* remove --locked (#533)

* Bump actions/labeler from 4.0.2 to 4.1.0 (#525)

* Provide a memory StateBackendClient (#523)

* Rename StateBackend::Standalone to StateBackend:Sled

* Copy utility files from sled crate since they cannot be used directly

* Provide a memory StateBackendClient

* Fix dashmap deadlock issue

* Fix for the comments


* only build docker images on rc tags (#535)

* docs: fix style in the Helm readme (#551)

* Fix Helm chart's image format (#550)

* Update datafusion requirement from 14.0.0 to 15.0.0 (#552)

* Update datafusion requirement from 14.0.0 to 15.0.0

* Fix UT

* Fix python

* Fix python

* Fix Python


* Make it concurrently to launch tasks to executors (#557)

* Make it concurrently to launch tasks to executors

* Refine for comments


* fix(ui): fix last seen (#562)

* Support Alibaba Cloud OSS with ObjectStore (#567)

* Fix cargo clippy (#571)


* Super minor spelling error (#573)

* Update env_logger requirement from 0.9 to 0.10 (#539)

Updates the requirements on [env_logger](https://github.com/rust-cli/env_logger) to permit the latest version.
- [Release notes](https://github.com/rust-cli/env_logger/releases)
- [Changelog](https://github.com/rust-cli/env_logger/blob/main/CHANGELOG.md)
- [Commits](https://github.com/rust-cli/env_logger/compare/v0.9.0...v0.10.0)

---
updated-dependencies:
- dependency-name: env_logger
  dependency-type: direct:production
...



* Update graphviz-rust requirement from 0.4.0 to 0.5.0 (#574)

Updates the requirements on [graphviz-rust](https://github.com/besok/graphviz-rust) to permit the latest version.
- [Release notes](https://github.com/besok/graphviz-rust/releases)
- [Changelog](https://github.com/besok/graphviz-rust/blob/master/CHANGELOG.md)
- [Commits](https://github.com/besok/graphviz-rust/commits)

---
updated-dependencies:
- dependency-name: graphviz-rust
  dependency-type: direct:production
...



* updated readme to contain correct versions of dependencies. (#580)

* Fix benchmark image link (#596)

* Add support for Azure (#599)

* Remove outdated script and use evergreen version of rust (#597)

* Remove outdated script and use evergreen version of rust

* Use debian protobuf

* Customize session builder

* Add setter for executor slots policy

* Construct Executor with functions

* Add queued and completed timestamps to successful job status

* Add public methods to SchedulerServer

* Public method for getting execution graph

* Public method for stage metrics

* Use node-level local limit (#20)

* Use node-level local limit

* serialize limit in shuffle writer

* Revert "Merge pull request #19 from coralogix/sc-5792"

This reverts commit 08140ef3f1ebe573ce9f0d2f3253422546c8a31a, reversing
changes made to a7f13844e44328203f803e5073cfff69f3981f4e.

* add log

* make sure we don't forget limit for shuffle writer

* update accum correctly and try to break early

* Check local limit accumulator before polling for more data

* fix build


* Add ClusterState trait

* Expose active job count

* Remove println

* Resubmit jobs when no resources available for scheduling

* Make parse_physical_expr public

* Reduce log spam

* Fix job submitted metric by ignoring resubmissions

* Record when job is queued in scheduler metrics (#28)

* Record when job is queueud in scheduler metrics

* add additional buckets for exec times

* fmt

* clippy

* tomlfmt

+++ ballista/core/src/execution_plans/shuffle_reader.rs
@@ -364,7 +364,7 @@ async fn fetch_partition_remote(
     // TODO for shuffle client connections, we should avoid creating new connections again and again.
     // And we should also avoid to keep alive too many connections for long time.
     let host = metadata.host.as_str();
-    let port = metadata.port as u16;
+    let port = metadata.port;
     let mut ballista_client =
         BallistaClient::try_new(host, port)
             .await
